requirements.txt
# =============================================================================
# CORE DEPENDENCIES
# =============================================================================

# LangGraph para orquestração de agentes
langgraph>=0.2.0

# LangChain core (dependência do LangGraph)
langchain>=0.1.0
langchain-core>=0.1.0

# HTTP requests
requests>=2.31.0

# Variáveis de ambiente
python-dotenv>=1.0.0

# =============================================================================
# UTILITIES
# =============================================================================

# Data e hora
python-dateutil>=2.8.2

# Validação de dados
pydantic>=2.0.0

# =============================================================================
# OPTIONAL: Database & Cache
# =============================================================================

# PostgreSQL (descomentar se usar banco de dados)
# psycopg2-binary>=2.9.9

# Redis (descomentar se usar cache)
# redis>=5.0.0

# =============================================================================
# OPTIONAL: Observability & Monitoring
# =============================================================================

# Prometheus metrics (descomentar se precisar)
# prometheus-client>=0.19.0

# Sentry error tracking (descomentar se precisar)
# sentry-sdk>=1.40.0

# =============================================================================
# DEVELOPMENT DEPENDENCIES
# =============================================================================

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Code quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Type stubs
types-requests>=2.31.0
types-python-dateutil>=2.8.0
